% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/neighbor_functions_dcemri.R
\name{indx}
\alias{indx}
\title{Get scalar index from two-dimensional index}
\usage{
indx(index, I, J)
}
\arguments{
\item{index}{two-dimensional index}

\item{I}{first dimension of array containing simulated measurements}

\item{J}{second dimension of array containing simulated measurements}
}
\value{
indexsc scalar index
}
\description{
Get scalar index from two-dimensional index
}

