% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/refit.R
\name{refit}
\alias{refit}
\title{Refit}
\usage{
refit(response, response_true = NULL, time, uniquekeps)
}
\arguments{
\item{response}{numeric vector containing the measurements at time points specified in argument time}

\item{response_true}{numeric vector containing the measurements without error at time points specified in vector time (for simulated measurements only)}

\item{time}{numeric vector containing the time points at which measurements are available}

\item{uniquekeps}{unique selected kep-estimates}
}
\value{
len: number of unique selected kep-estimates

sse: residual sum of squares

sse_true: residual sum of squares for measurements without error (for simulated measurements only)

kk_esti: kep- and Ktrans-estimates
}
\description{
Refit
}

