% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boosting_spatial.R
\name{spatboost}
\alias{spatboost}
\title{Spatial boosting}
\usage{
spatboost(response, offset, niter, nu, time, factor, M, SSEbefore, ALLKEP,
  whichnb, lambda, kep_penal)
}
\arguments{
\item{response}{numeric vector containing the measurements at time points specified in argument time}

\item{offset}{offset}

\item{niter}{number of iterations}

\item{nu}{step length factor}

\item{time}{numeric vector containing the time points at which measurements are available}

\item{factor}{factor by which the kep-values should differ at least}

\item{M}{current iteration}

\item{SSEbefore}{residual sum of squares in the last iteration}

\item{ALLKEP}{kep-estimates accepted in previous iterations}

\item{whichnb}{index for the neighbors of the voxel}

\item{lambda}{penalization parameter for spatial regularization}

\item{kep_penal}{kep-estimates of the neighbors of the voxel}
}
\value{
response: numeric vector containing the measurements at time points specified in argument time

coeff: kep- and Ktrans-estimates

fit: numeric vector containing the current fit

u: numeric vector containing the current residuals

DF: degrees of freedom

IC: value of the information criterion (BIC)

SSE: residual sum of squares

m: current iteration
}
\description{
Spatial boosting
}

