% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dcemriboost-package.r
\docType{package}
\name{dcemriboost}
\alias{dcemriboost}
\alias{dcemriboost-package}
\title{dcemriboost - Boosting for DCEMRI data.}
\description{
With dcemriboost voxelwise and spatially regularized boosting for DCEMRI data can be performed.
}
\details{
Implemented methods are

boostvw: voxelwise boosting

spatboost: spatially regularized boosting

refit: Refit for voxelwise and spatially regularized boosting
}
\author{
Martina Feilke \email{martina.feilke@stat.uni-muenchen.de}

Maintainer: Martina Feilke \email{martina.feilke@stat.uni-muenchen.de}
}

